requirements.txt
Flask==3.0.2
Pillow==10.2.0
python-magic==0.4.27 